Full rework of Key storage.  Node keys are stored locally using user specific keys.  Personal keys are now stored in the network.  For security they are split into 3 which are shuffled and encrypted in a random order with each other. We store the first of the Alpha's first bit to track it and mix it  with verification bits that are used to identify the next, all 3 parts are mixed with fake parts so the only way to know the correct next part is to successfully guess the preceding part.  (Alpha<Beta<Delta) if you manage to decrypt all 3 parts they would still need to be reassembled in the correct order.
The Alpha is encrypted with the users password and pin as salt for security. When a user signs into a node with their id, password and pin,  their encrpytedKeyParts is downloaded and if it is successfully Decrypted then the key is stored securely in memory and the user can use it and is signed in.

This allowed to Remove the need for Service accounts and CNG containers,  This will make it easier to build and accommodate other languages in the future.